requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.2
werkzeug==2.0.1
python-docx==0.8.11
PyPDF2==3.0.1
pandas==1.5.3
python-magic==0.4.27
nltk==3.8.1